requirements.txt
langchain>=0.1.0
ragas>=0.1.0
pandas>=2.0.0
openai>=1.0.0
requests>=2.25.0
tiktoken>=0.5.0
datasets>=2.0.0
scikit-learn>=1.0.0
jupyter>=1.0.0
python-dotenv>=1.0.0
